Move blocks query to a `getBlocks` query function
